device_tags_python/requirements.txt
click==8.0.4
balena-sdk==15.0.2
Flask==2.0.3
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
Werkzeug==2.0.3
zipp==3.6.0
